client_protocol: refactor s3 config into  crd::s3 and config::s3 to make it reusable
