+++ rust/operator-binary/src/config/mod.rs
@@ -1,1 +1,3 @@
+pub mod client_protocol;
 pub mod jvm;
+pub mod s3;
